terraform/aws: enable vpc endpoint for internal networks
By configuring this VPC enpoint it will route traffic from the internal
subnets to S3 via a this "virtual" endpoint. This will offload traffic
from the NAT instance, so it will be less of a bottleneck for the
cluster when there are a lot of interactions between CF and S3.